add delete? create server method method

+++ airport/src/main/java/com/ra/airport/servlet/handler/DeleteAirportHandler.java
@@ -1,14 +1,12 @@
 package com.ra.airport.servlet.handler;
 
-import com.google.gson.Gson;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.ra.airport.entity.Airport;
 import com.ra.airport.repository.exception.AirPortDaoException;
 import com.ra.airport.service.AirPortService;
-import com.ra.airport.service.AirportServiceImpl;
+import org.apache.logging.log4j.util.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -18,27 +16,29 @@ public class DeleteAirportHandler implements ServletHandler {
     private final transient AirPortService<Airport> airportService;
 
     @Autowired
-    public DeleteAirportHandler(final AirportServiceImpl airportService) {
+    public DeleteAirportHandler(final AirPortService<Airport> airportService) {
         this.airportService = airportService;
     }
 
     @Override
     public void post(final HttpServletRequest request, final HttpServletResponse response) throws AirPortDaoException {
-        final String jsonAirport = (String) request.getAttribute("airport");
-        final Gson gson = new Gson();
-        final Airport airport = gson.fromJson(jsonAirport, Airport.class);
-        airportService.delete(airport);
-        request.setAttribute("jspPath", "deleteAirport");
+        delete(request, response);
+
     }
 
     @Override
     public void get(final HttpServletRequest request, final HttpServletResponse response) throws AirPortDaoException {
-        post(request, response);
+        delete(request, response);
     }
 
     @Override
     public void delete(final HttpServletRequest request, final HttpServletResponse response) throws AirPortDaoException {
-
+        final String airportId = request.getParameter("apId");
+        if (Strings.isNotBlank(airportId)) {
+            final var airport = new Airport();
+            airport.setApId(Integer.parseInt(airportId));
+            airportService.delete(airport);
+        }
     }
 
     @Override
